+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import SimpleSlider from './components/Carousel';
+import SimpleSlider from './components/carousel/Carousel';
 
 class App extends Component {
   render() {
